refactor: Appointment 서비스에 CQS(Command Query Separation) 패턴 적용
