+++ src/main/java/com/codelion/animalcare/domain/appointment/service/AppointmentQueryService.java
@@ -1,6 +1,7 @@
 package com.codelion.animalcare.domain.appointment.service;
 
 import com.codelion.animalcare.domain.appointment.dto.AppointmentDto;
+import com.codelion.animalcare.domain.appointment.dto.AppointmentModifyDto;
 import com.codelion.animalcare.domain.appointment.dto.LoadMyPageDoctorAppointment;
 import com.codelion.animalcare.domain.appointment.entity.Appointment;
 import com.codelion.animalcare.domain.appointment.repository.AppointmentRepository;
@@ -12,7 +13,10 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -57,6 +61,20 @@ public class AppointmentQueryService {
         return result;
     }
 
+    /**
+     * 의사가 해당 날짜에 예약 되어있는 시간을 출력함.
+     * @param date
+     * @param doctorId
+     * @return
+     */
+    public List<LocalDateTime> findDateTimesByDateAndDoctor(LocalDate date, Long doctorId){
+        // UTC로 검색하기 위해 Java.sql.Date 대신 LocalDate 사용
+        LocalDateTime utcDateTimeFront = date.atStartOfDay();
+        LocalDateTime utcDateTimeEnd = date.atStartOfDay().plusDays(1);
+        System.out.println(utcDateTimeFront + " " + utcDateTimeEnd);
+        return appointmentRepository.findDateTimesByDateAndDoctor(utcDateTimeFront, utcDateTimeEnd, doctorId);
+    }
+
 
 
 
@@ -82,5 +100,22 @@ public class AppointmentQueryService {
     private Member findMember(String email) {
         return memberRepository.findByEmail(email)
                 .orElseThrow(() -> new RuntimeException("Member " + email + "is not found."));
+    }
+
+
+    public LoadMyPageDoctorAppointment.ResponseDto findById(long appointmentId) {
+        Appointment appointment = appointmentRepository
+                .findByAppointmentId(appointmentId)
+                .orElseThrow(() -> new RuntimeException("Appointment id " + appointmentId + " is not found."));
+
+        return new LoadMyPageDoctorAppointment.ResponseDto(appointment);
+    }
+
+    public Optional<AppointmentModifyDto> findAppointmentModifyDtoById(Long appointmentId) {
+
+        Optional<Appointment> appointmentOptional = appointmentRepository.findByAppointmentId(appointmentId);
+        Optional<AppointmentModifyDto> appointmentModifyDto = appointmentOptional.map(o -> new AppointmentModifyDto(o));
+
+        return appointmentModifyDto;
     }
 }
