+++ src/test/java/com/codelion/animalcare/domain/appointment/AppointmentServiceTests.java
@@ -1,0 +1,10 @@
+package com.codelion.animalcare.domain.appointment;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class AppointmentServiceTests {
+
+
+
+}
